ISPN-6905 Remove VersionAwaremarshaller
