+++ core/src/test/java/org/infinispan/marshall/multiversion/MultiPojoVersionMarshallTest.java
@@ -5,11 +5,10 @@ import javassist.ClassPool;
 import javassist.CtClass;
 import javassist.CtField;
 import javassist.CtMethod;
-import org.infinispan.commons.marshall.Marshaller;
+import org.infinispan.commons.marshall.StreamingMarshaller;
 import org.infinispan.configuration.cache.ConfigurationBuilder;
 import org.infinispan.configuration.global.GlobalConfigurationBuilder;
 import org.infinispan.manager.EmbeddedCacheManager;
-import org.infinispan.commons.marshall.AbstractDelegatingMarshaller;
 import org.infinispan.commons.marshall.Externalizer;
 import org.infinispan.commons.marshall.SerializeWith;
 import org.infinispan.commons.util.Util;
@@ -57,7 +56,7 @@ public class MultiPojoVersionMarshallTest extends AbstractInfinispanTest {
    private static final String HOUSE = BASE + "House";
    private static final String HOUSE_EXT = HOUSE + "Externalizer";
 
-   private AbstractDelegatingMarshaller marshaller;
+   private StreamingMarshaller marshaller;
    private EmbeddedCacheManager cm;
 
    @BeforeClass
@@ -355,7 +354,7 @@ public class MultiPojoVersionMarshallTest extends AbstractInfinispanTest {
       }
    }
 
-   private AbstractDelegatingMarshaller getMarshaller() {
+   private StreamingMarshaller getMarshaller() {
       return marshaller;
    }
