+++ core/src/test/resources/configs/unified/8.0.xml
@@ -32,7 +32,7 @@
                     jndi-name="java:global/infinispan/maximal"  state-transfer-executor="infinispan-state-transfer" listener-executor="infinispan-listener" replication-queue-executor="infinispan-repl-queue"
                     persistence-executor="infinispan-cached" module="org.infinispan" statistics="true" shutdown-hook="DONT_REGISTER">
       <transport cluster="maximal-cluster" executor="infinispan-transport" total-order-executor="infinispan-cached" remote-command-executor="infinispan-cached" lock-timeout="120000" stack="tcp" node-name="a-node" machine="a" rack="b" site="c"/>
-      <serialization marshaller="org.infinispan.marshall.core.VersionAwareMarshaller" version="1.0">
+      <serialization marshaller="org.infinispan.marshall.TestObjectStreamMarshaller" version="1.0">
          <advanced-externalizer class="org.infinispan.marshall.AdvancedExternalizerTest$IdViaConfigObj$Externalizer" id="9001" />
          <advanced-externalizer class="org.infinispan.marshall.AdvancedExternalizerTest$IdViaAnnotationObj$Externalizer" id="9002" />
       </serialization>
